+++ TP2_Bieres_DUMAS/src/tp2_bieres_dumas/TP2_Bieres_DUMAS.java
@@ -9,9 +9,9 @@ public class TP2_Bieres_DUMAS {
     public static void main(String[] args) {
 
  
-        BouteilleBiere uneBiere = new BouteilleBiere("Cuvée des trolls", 7.0 , "Dubuisson");
-        BouteilleBiere autreBiere = new BouteilleBiere("Leffe", 6.6 , "Abbaye de Leffe");
-        BouteilleBiere troisiemeBiere = new BouteilleBiere("Bière", 5.0, "Brasserie3");
+        BouteilleBiere uneBiere = new BouteilleBiere("Cuvée des trolls", (float) 7.0, "Dubuisson");
+        BouteilleBiere autreBiere = new BouteilleBiere("Leffe", (float) 6.6, "Abbaye de Leffe");
+        BouteilleBiere troisiemeBiere = new BouteilleBiere("Bière", (float) 5.0, "Brasserie3");
 
         uneBiere.lireEtiquette();
         autreBiere.lireEtiquette();
@@ -24,4 +24,3 @@ public class TP2_Bieres_DUMAS {
         System.out.println(troisiemeBiere);
     }
 }
-
